Service, repos refactor
