+++ src/repositories/survey_repository.py
@@ -14,19 +14,6 @@ class SurveyRepository:
             print(e)
             return False
 
-    def add_new_survey(self, surveyname, teacher_id):
-        try:
-            sql = "INSERT INTO surveys (surveyname, teacher_id, min_choices, closed, results_saved) VALUES (:surveyname, :teacher_id, :min_choices, :closed, :results_saved) RETURNING id"
-            result = db.session.execute(text(sql), {"surveyname":surveyname, "teacher_id":teacher_id, "min_choices":10, "closed":False, "results_saved":False})
-            db.session.commit()
-            survey = result.fetchone()[0]
-            if not survey:
-                return False
-            return survey
-        except Exception as e: # pylint: disable=W0718
-            print(e)
-            return False
-
     def survey_name_exists(self, surveyname, teacher_id):
         try:
             sql = "SELECT id FROM surveys WHERE (surveyname=:surveyname AND teacher_id=:teacher_id AND closed=False)"
@@ -104,7 +91,6 @@ class SurveyRepository:
             print(e)
             return False
 
-    
     def create_new_survey(self, surveyname, user_id, min_choices =1):
         '''
         Creates a new survey, updates just surveys table
@@ -116,39 +102,5 @@ class SurveyRepository:
         result = db.session.execute(text(sql), {"surveyname":surveyname, "teacher_id":user_id, "min_choices":min_choices, "closed":False, "saved":False})
         db.session.commit()
         return result.fetchone()[0]
-
-    def create_new_survey_choice(self, survey_id, name, seats):
-        '''
-        Adds a new choice to existing survey, updates just survey_choices table
-        RETURNS created choice's id
-        '''
-        sql = "INSERT INTO survey_choices (survey_id, name, max_spaces)"\
-              " VALUES (:survey_id, :name, :max_spaces) RETURNING id"
-        result = db.session.execute(text(sql), {"survey_id":survey_id, "name":name, "max_spaces":seats})
-        db.session.commit()
-        return result.fetchone()[0]
-
-    def create_new_choice_info(self, choice_id, info_key, info_value):
-        '''
-        Adds an additional to existing survey choice, updates choice_infos table
-        '''
-        sql = "INSERT INTO choice_infos (choice_id, info_key, info_value)"\
-              " VALUES (:c_id, :i_key, :i_value)"
-        result = db.session.execute(text(sql), {"c_id":choice_id, "i_key":info_key, "i_value":info_value})
-        db.session.commit()
-
-    def get_choice_additional_infos(self, choice_id):
-        '''
-        Gets a list of key-value pairs based on choice_id from choice_infos tables
-        '''
-
-        sql = "SELECT info_key, info_value FROM choice_infos WHERE choice_id=:choice_id"
-        result = db.session.execute(text(sql), {"choice_id":choice_id})
-        return result.fetchall()
-
-    def get_choice_name_and_spaces(self, choice_id):
-        sql = "SELECT name, max_spaces FROM survey_choices WHERE id=:id"
-        result = db.session.execute(text(sql), {"id":choice_id})
-        return result.fetchone()
 
 survey_repository = SurveyRepository()
