+++ src/tools/parsers.py
@@ -1,5 +1,7 @@
 from src import db
 from src.repositories.survey_repository import survey_repository
+from src.repositories.survey_choices_repository import survey_choices_repository
+from src.repositories.choice_infos_repository import choice_infos_repository
 from sqlalchemy import text
 
 def clear_database():
@@ -47,12 +49,12 @@ def parser_elomake_csv(file, survey_name, user_id):
         name = ''.join(c for c in temp[2] if c.isprintable())
         spaces = ''.join(c for c in temp[3] if c.isprintable()).strip('"')
 
-        choice_id = survey_repository.create_new_survey_choice(survey_id, name, int(spaces))
+        choice_id = survey_choices_repository.create_new_survey_choice(survey_id, name, int(spaces))
 
         i = 4
         while i < col_count:
             temp_string = ''.join(c for c in temp[i] if c.isprintable())
-            survey_repository.create_new_choice_info(choice_id, info_headers[i], temp_string.strip('"'))
+            choice_infos_repository.create_new_choice_info(choice_id, info_headers[i], temp_string.strip('"'))
             i += 1
 
         index += 1
@@ -76,8 +78,8 @@ def parser_manual(survey_choices, survey_name, user_id):
             if count == 1:
                 spaces = choice[pair]
                 count += 1
-                choice_id = survey_repository.create_new_survey_choice(survey_id, name, spaces)
+                choice_id = survey_choices_repository.create_new_survey_choice(survey_id, name, spaces)
                 continue
 
-            survey_repository.create_new_choice_info(choice_id, pair, choice[pair])
+            choice_infos_repository.create_new_choice_info(choice_id, pair, choice[pair])
             count += 1
